server/requirements.txt
Flask==1.1.2
gunicorn==20.0.4
nltk==3.6.2
openai==0.10.2
newspaper3k>=0.2.8
python-dotenv==0.19.0
requests>=2.25.1
